feat: adding features for saving, general  searching and by id

+++ desafio-sinapses-marcos/desafio/projeto base/sinapses/src/main/java/br/com/zaffari/sinapses/repository/SinapseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.zaffari.sinapses.model.Sinapse;
 
 @Repository
-public interface SinapseRepository extends JpaRepository<Sinapse, Integer>{
+public interface SinapseRepository extends JpaRepository<Sinapse, Long>{
 
 }
